feat: fixes
