+++ packages/relayer/eml_templates/guardian_not_set.html
@@ -104,7 +104,7 @@
                   "
                 >
                 <img
-                src="https://storage.googleapis.com/eml-templates-assets/img/logoEmailWallet.png"
+                src="https://storage.googleapis.com/eml-templates-assets/img/logo.png"
                 alt="Email Wallet"
                 style="
                   width: 75%;
