+++ packages/utils/src/regex.rs
@@ -1,5 +1,4 @@
 use neon::prelude::*;
-use neon::prelude::*;
 pub use zk_regex_apis::extract_substrs::*;
 pub use zk_regex_apis::padding::*;
